Scale on window resize, and add code editor.
Scale the game manually on resize (and at other times) so that none of
the screen will be chopped off.

Add the Ace editor so users can edit their code within the browser
rather than in Notepad or some other bare-bones editor.

This commit means the assets and game dimensions will need to be resized
to 8:9 rather than 16:9.